Allow much more customization in ChatViewController
Closes #773
